Additional updates in code base to compiler with the new size_t return of strlen in x64.

+++ Pulp/3DEngine/MaterialManager.cpp
@@ -45,7 +45,7 @@ namespace {
 
 	bool strToVec4(const char* pStr, Vec4f& vecOut)
 	{
-		uint32_t len = core::strUtil::strlen(pStr);
+		size_t len = core::strUtil::strlen(pStr);
 		if (len > 128)
 			return false;
